format_percentage

rounds percentage values while allowing the option to `trim_trailing_zeros`

Usage

format_percentage(value, round_digits, trim_trailing_zeros)

Arguments

value

numeric; percentage value to be rounded

round_digits

integer; number of digits past the decimal to round to

trim_trailing_zeros

logical; trim trailing zeros or not in returned value.